Moved 2D XY frame with gravity load from benchmark module to its own module
